switched it back PhysVecRole back in Molecule
The atomgrouprole has a total_mass method already. It may be a good idea
to switch, but needs some more thinking than I can do tonight.  Adding
the PhysVecRole into the AtomGroup opens all the groups to have arrays of
forces and coordinates (COM, etc).  Could be nice, but would this be too much?